Package examples and remove `dev-dependencies`
