+++ Cargo.toml
@@ -29,6 +29,18 @@ members = [
     "web",
     "wgpu",
     "winit",
+    "examples/bezier_tool",
+    "examples/counter",
+    "examples/custom_widget",
+    "examples/events",
+    "examples/geometry",
+    "examples/pokedex",
+    "examples/progress_bar",
+    "examples/stopwatch",
+    "examples/styling",
+    "examples/svg",
+    "examples/todos",
+    "examples/tour",
 ]
 
 [target.'cfg(not(target_arch = "wasm32"))'.dependencies]
@@ -37,20 +49,3 @@ iced_wgpu = { version = "0.1.0", path = "wgpu" }
 
 [target.'cfg(target_arch = "wasm32")'.dependencies]
 iced_web = { version = "0.1.0", path = "web" }
-
-[dev-dependencies]
-iced_native = { version = "0.1", path = "./native" }
-iced_wgpu = { version = "0.1", path = "./wgpu" }
-iced_futures = { version = "0.1.0-alpha", path = "./futures", features = ["async-std"] }
-env_logger = "0.7"
-serde = { version = "1.0", features = ["derive"] }
-serde_json = "1.0"
-directories = "2.0"
-futures = "0.3"
-async-std = { version = "1.3", features = ["unstable"] }
-surf = "1.0"
-rand = "0.7"
-lyon = "0.15"
-
-[target.'cfg(target_arch = "wasm32")'.dev-dependencies]
-wasm-bindgen = "0.2.51"
