+++ examples/counter/src/main.rs
