+++ examples/events/src/main.rs
