+++ examples/geometry/src/main.rs
