+++ examples/progress_bar/src/main.rs
