+++ examples/tour/images/ferris.png
