+++ examples/tour/src/main.rs
@@ -681,10 +681,10 @@ fn ferris<'a>(width: u16) -> Container<'a, StepMessage> {
         // This should go away once we unify resource loading on native
         // platforms
         if cfg!(target_arch = "wasm32") {
-            Image::new("resources/ferris.png")
+            Image::new("images/ferris.png")
         } else {
             Image::new(format!(
-                "{}/examples/resources/ferris.png",
+                "{}/images/ferris.png",
                 env!("CARGO_MANIFEST_DIR")
             ))
         }
